// src/services/types/platform/assess_prioritize.ts
export type AssessPrioritizeApiResponse = {
  _id: string
  __v: number
  assess_and_prioritize: {
    section1: {
      title1_en: string
      title1_fa: string
      title2_en: string
      title2_fa: string
      text1_en: string
      text1_fa: string
      button1_en: string
      button1_fa: string
    }
    section2: {
      title1_en: string
      title1_fa: string
      text1_en: string
      text1_fa: string
    }
    section3: {
      image1: string
    }
    section4: {
      title1_en: string
      title1_fa: string
      title2_en: string
      title2_fa: string
      text1_en: string
      text1_fa: string
      button1_en: string
      button1_fa: string
      image1: string
    }
    section5: {
      title1_en: string
      title1_fa: string
      text1_en: string
      text1_fa: string
      image1: string
    }
    section6: {
      title1_en: string
      title1_fa: string
      title2_en: string
      title2_fa: string
      text1_en: string
      text1_fa: string
      image1: string
    }
    section7: {
      text1_en: string
      text1_fa: string
      text2_en: string
      text2_fa: string
    }
    section8: {
      text1_en: string
      text1_fa: string
      text2_en: string
      text2_fa: string
      button1_en: string
      button1_fa: string
    }
  }
}